# fd-microservice/openapi_spec.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/admin/time-travel":  {
                                                 "post":  {
                                                              "tags":  [
                                                                           "admin-controller"
                                                                       ],
                                                              "operationId":  "timeTravel",
                                                              "requestBody":  {
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/TimeTravelRequest"
                                                                                                                                      }
                                                                                                                       }
                                                                                              },
                                                                                  "required":  true
                                                                              },
                                                              "responses":  {
                                                                                "404":  {
                                                                                            "description":  "Not Found",
                                                                                            "content":  {
                                                                                                            "*/*":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                                        }
                                                                                        },
                                                                                "400":  {
                                                                                            "description":  "Bad Request",
                                                                                            "content":  {
                                                                                                            "*/*":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                                        }
                                                                                        },
                                                                                "500":  {
                                                                                            "description":  "Internal Server Error",
                                                                                            "content":  {
                                                                                                            "*/*":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                                        }
                                                                                        },
                                                                                "403":  {
                                                                                            "description":  "Forbidden",
                                                                                            "content":  {
                                                                                                            "*/*":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                                        }
                                                                                        },
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  {
                                                                                                            "*/*":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          }
                                             },
                  "/api/admin/batch/statement-generation":  {
                                                                "post":  {
                                                                             "tags":  [
                                                                                          "admin-controller"
                                                                                      ],
                                                                             "operationId":  "triggerStatementGeneration",
                                                                             "responses":  {
                                                                                               "404":  {
                                                                                                           "description":  "Not Found",
                                                                                                           "content":  {
                                                                                                                           "*/*":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                                       }
                                                                                                       },
                                                                                               "400":  {
                                                                                                           "description":  "Bad Request",
                                                                                                           "content":  {
                                                                                                                           "*/*":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                                       }
                                                                                                       },
                                                                                               "500":  {
                                                                                                           "description":  "Internal Server Error",
                                                                                                           "content":  {
                                                                                                                           "*/*":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                                       }
                                                                                                       },
                                                                                               "403":  {
                                                                                                           "description":  "Forbidden",
                                                                                                           "content":  {
                                                                                                                           "*/*":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                                       }
                                                                                                       },
                                                                                               "200":  {
                                                                                                           "description":  "OK",
                                                                                                           "content":  {
                                                                                                                           "*/*":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                                       }
                                                                                                       }
                                                                                           }
                                                                         }
                                                            },
                  "/api/admin/batch/maturity-processing":  {
                                                               "post":  {
                                                                            "tags":  [
                                                                                         "admin-controller"
                                                                                     ],
                                                                            "operationId":  "triggerMaturityProcessing",
                                                                            "responses":  {
                                                                                              "404":  {
                                                                                                          "description":  "Not Found",
                                                                                                          "content":  {
                                                                                                                          "*/*":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                      }
                                                                                                      },
                                                                                              "400":  {
                                                                                                          "description":  "Bad Request",
                                                                                                          "content":  {
                                                                                                                          "*/*":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                      }
                                                                                                      },
                                                                                              "500":  {
                                                                                                          "description":  "Internal Server Error",
                                                                                                          "content":  {
                                                                                                                          "*/*":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                      }
                                                                                                      },
                                                                                              "403":  {
                                                                                                          "description":  "Forbidden",
                                                                                                          "content":  {
                                                                                                                          "*/*":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                      }
                                                                                                      },
                                                                                              "200":  {
                                                                                                          "description":  "OK",
                                                                                                          "content":  {
                                                                                                                          "*/*":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                      }
                                                                                                      }
                                                                                          }
                                                                        }
                                                           },
...
              },
    "components":  {
                       "schemas":  {
                                       "ApiResponse":  {
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "success":  {
                                                                                              "type":  "boolean"
                                                                                          },
                                                                              "message":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                              "data":  {
                                                                                           "type":  "object"
                                                                                       },
                                                                              "timestamp":  {
                                                                                                "type":  "string",
                                                                                                "format":  "date-time"
                                                                                            }
                                                                          }
                                                       },
...
                                       "TimeTravelRequest":  {
                                                                 "required":  [
                                                                                  "targetDate"
                                                                              ],
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "targetDate":  {
                                                                                                       "type":  "string",
                                                                                                       "format":  "date"
                                                                                                   },
                                                                                    "operation":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                }
                                                             },
...
                                   }
                   }
}